update ROS/BulletROS + add simulator examples

+++ pyrobolearn/simulators/bullet_ros.py
@@ -27,7 +27,8 @@ import os
 import subprocess
 import psutil
 import signal
-import rospy
+import importlib
+import inspect
 
 # from pyrobolearn.simulators.simulator import Simulator
 from pyrobolearn.simulators.bullet import Bullet
@@ -54,7 +55,7 @@ class BulletROS(Bullet):  # , ROS):
     sensors, actuators, and forces, to map the real world to the simulated one, etc.
     """
 
-    def __init__(self, render=True, subscribe=False, publish=False, ros_master_uri=11311, **kwargs):
+    def __init__(self, render=True, subscribe=False, publish=False, teleoperate=False, ros_master_uri=11311, **kwargs):
         """
         Initialize the Bullet-ROS simulator.
 
@@ -80,10 +81,13 @@ class BulletROS(Bullet):  # , ROS):
             # subprocess.Popen("roscore", env=self.env)
             self.roscore = subprocess.Popen(["roscore", "-p", str(ros_master_uri)], env=self.env,
                                             preexec_fn=os.setsid)  # , shell=True)
+        else:
+            print('ROS core has already been initialized.')
 
         # set variables
         self.subscribe = subscribe
         self.publish = publish
+        self.teleoperate = teleoperate
 
         # remember each publisher/subscriber
         self.subscribers = {}
@@ -98,7 +102,11 @@ class BulletROS(Bullet):  # , ROS):
         # delete each publishers
 
         # delete ROS
-        os.killpg(os.getpgid(self.roscore.pid), signal.SIGTERM)
+        if self.roscore is not None:
+            os.killpg(os.getpgid(self.roscore.pid), signal.SIGTERM)
+
+        # call parent destructor
+        super(BulletROS, self).close()
 
     @property
     def is_subscribing(self):
@@ -160,19 +168,42 @@ class BulletROS(Bullet):  # , ROS):
                                                use_maximal_coordinates=use_maximal_coordinates,
                                                use_fixed_base=use_fixed_base, flags=flags, scale=scale)
         # get path to directory of urdf
-        path = os.path.dirname(filename)
-        robot_directory_name = path.split('/')[-1]
-        path = path + '/../../ros/' + robot_directory_name + '/'
+        path = os.path.dirname(os.path.abspath(filename))               # /path/to/pyrobolearn/robots/urdfs/<robot>/
+        robot_directory_name = path.split('/')[-1]                      # <robot>
+        # path = path + '/../../ros/' + robot_directory_name + '/'        # /path/to/pyrobolearn/robots/ros/<robot>/
 
         # check if valid robot directory
+        # if os.path.isdir(path):
+        robot_path = '/'.join(path.split('/')[-5:-2])
+        if robot_path == 'pyrobolearn/robots/urdfs':
 
-        # load subscriber in simulator
-        if self.subscribe:
-            pass
+            # get corresponding subscriber/publisher
+            def check_ros(name, dictionary, id_):
+                # TODO: do I really need a class for each robot? Can I not just use RobotPublisher?
+                # if os.path.isfile(path + name + '.py'):
+                #     module = importlib.import_module('pyrobolearn.robots.ros.' + robot_directory_name + '.' + name)
+                #     classes = inspect.getmembers(module, inspect.isclass)  # list of (name, class)
+                #     length = len(name)
+                #     robot_name = ''.join(robot_directory_name.split('_'))
+                #
+                #     # go through each class and get the :attr:`name` corresponding to the robot and add it to the
+                #     # given :attr:`dictionary`
+                #     for name, cls in classes:
+                #         if name[:-length].lower() == robot_name:
+                #             dictionary[id_] = cls(id_=id_)
+                #             break
+                module = importlib.import_module('pyrobolearn.robots.ros.' + name)
+                classes = dict(inspect.getmembers(module, inspect.isclass))
+                cls = classes['Robot' + name.capitalize()]
+                dictionary[id_] = cls(name=robot_directory_name, id_=id_)
 
-        # load publisher in simulator
-        if self.publish:
-            pass
+            # load subscriber in simulator
+            if self.subscribe:
+                check_ros('subscriber', self.subscribers, id_)
+
+            # load publisher in simulator
+            if self.publish:
+                check_ros('publisher', self.publishers, id_)
 
         return id_
 
@@ -191,10 +222,16 @@ class BulletROS(Bullet):  # , ROS):
                 np.float[N]: joint positions [rad]
         """
         if body_id in self.subscribers:
-            q = self.subscribers[body_id].get_joint_positions[joint_ids]
-            super(BulletROS, self).set_joint_positions(body_id=body_id, joint_ids=joint_ids, positions=q)  # or reset?
+            q = self.subscribers[body_id].get_joint_positions(joint_ids)
+            if len(q) == len(joint_ids):
+                super(BulletROS, self).set_joint_positions(body_id=body_id, joint_ids=joint_ids, positions=q)  # reset?
+            else:  # if failed to get the joint positions from the subscriber
+                q = super(BulletROS, self).get_joint_positions(body_id=body_id, joint_ids=joint_ids)
         else:
             q = super(BulletROS, self).get_joint_positions(body_id=body_id, joint_ids=joint_ids)
+            if self.teleoperate and body_id in self.publishers:
+                self.publishers[body_id].set_joint_positions(joint_ids, q)
+                self.publishers[body_id].publish('joint_states')
         return q
 
     def set_joint_positions(self, body_id, joint_ids, positions, velocities=None, kps=None, kds=None, forces=None):
@@ -213,6 +250,7 @@ class BulletROS(Bullet):  # , ROS):
         super(BulletROS, self).set_joint_positions(body_id, joint_ids, positions, velocities, kps, kds, forces)
         if body_id in self.publishers:
             self.publishers[body_id].set_joint_positions(joint_ids, positions)
+            self.publishers[body_id].publish('joint_states')
 
     def get_joint_velocities(self, body_id, joint_ids):
         """
@@ -228,7 +266,18 @@ class BulletROS(Bullet):  # , ROS):
             if multiple joints:
                 np.float[N]: joint velocities [rad/s]
         """
-        pass
+        if body_id in self.subscribers:
+            dq = self.subscribers[body_id].get_joint_velocities(joint_ids)
+            if len(dq) == len(joint_ids):
+                super(BulletROS, self).set_joint_velocities(body_id=body_id, joint_ids=joint_ids, velocities=dq)
+            else:  # if failed to get the joint velocities from the subscriber
+                dq = super(BulletROS, self).get_joint_velocities(body_id=body_id, joint_ids=joint_ids)
+        else:
+            dq = super(BulletROS, self).get_joint_velocities(body_id=body_id, joint_ids=joint_ids)
+            if self.teleoperate and body_id in self.publishers:
+                self.publishers[body_id].set_joint_velocities(joint_ids, dq)
+                self.publishers[body_id].publish('joint_states')
+        return dq
 
     def set_joint_velocities(self, body_id, joint_ids, velocities, max_force=None):
         """
@@ -240,7 +289,10 @@ class BulletROS(Bullet):  # , ROS):
             velocities (float, np.float[N]): desired velocity, or list of desired velocities [rad/s]
             max_force (None, float, np.float[N]): maximum motor forces/torques
         """
-        pass
+        super(BulletROS, self).set_joint_velocities(body_id, joint_ids, velocities, max_force)
+        if body_id in self.publishers:
+            self.publishers[body_id].set_joint_velocities(joint_ids, velocities)
+            self.publishers[body_id].publish('joint_states')
 
     def get_joint_torques(self, body_id, joint_ids):
         """
@@ -258,7 +310,18 @@ class BulletROS(Bullet):  # , ROS):
             if multiple joints:
                 np.float[N]: torques associated to the given joints [Nm]
         """
-        pass
+        if body_id in self.subscribers:
+            torques = self.subscribers[body_id].get_joint_torques(joint_ids)
+            if len(torques) == len(joint_ids):
+                super(BulletROS, self).set_joint_torques(body_id=body_id, joint_ids=joint_ids, torques=torques)
+            else:  # if failed to get the joint torques from the subscriber
+                torques = super(BulletROS, self).get_joint_torques(body_id=body_id, joint_ids=joint_ids)
+        else:
+            torques = super(BulletROS, self).get_joint_torques(body_id=body_id, joint_ids=joint_ids)
+            if self.teleoperate and body_id in self.publishers:
+                self.publishers[body_id].set_joint_torques(joint_ids, torques)
+                self.publishers[body_id].publish('joint_states')
+        return torques
 
     def set_joint_torques(self, body_id, joint_ids, torques):
         """
@@ -269,7 +332,10 @@ class BulletROS(Bullet):  # , ROS):
             joint_ids (int, list of int): joint id, or list of joint ids.
             torques (float, list of float): desired torque(s) to apply to the joint(s) [N].
         """
-        pass
+        super(BulletROS, self).set_joint_torques(body_id, joint_ids, torques)
+        if body_id in self.publishers:
+            self.publishers[body_id].set_joint_torques(joint_ids, torques)
+            self.publishers[body_id].publish('joint_states')
 
 
 # Test
